worker/requirements.txt
confluent-kafka>=2.0.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
cryptography>=42.0.0
httpx>=0.27.0
orjson>=3.9
python-dotenv>=1.0.0
